Rename default path selector to type name
