+++ iroh/src/socket/remote_map.rs
@@ -20,14 +20,16 @@ pub use self::remote_state::{
     TransportAddrUsage,
 };
 // Public re-exports used by `endpoint::transports` behind `unstable-custom-transports`.
-// Without that feature the types are still reachable in-crate (path_selector uses
-// `PathSelectionContext` in the trait signature), but they don't reach the public crate
+// Without that feature the types are still reachable in-crate (the default selector uses
+// `PathSelectionContext` in its trait signature), but they don't reach the public crate
 // boundary, so silence the resulting `unreachable_pub` deny.
 #[cfg_attr(
     not(feature = "unstable-custom-transports"),
     allow(unreachable_pub, unused_imports)
 )]
-pub use self::remote_state::{PathSelectionContext, PathSelectionData};
+pub use self::remote_state::{
+    PathSelection, PathSelectionContext, PathSelectionData, PathSelector,
+};
 use super::{
     DirectAddr, Metrics as SocketMetrics,
     mapped_addrs::{
@@ -40,7 +42,6 @@ use crate::{
     socket::{
         RemoteStateActorStoppedError,
         concurrent_read_map::{ConcurrentReadMap, ReadOnlyMap},
-        path_selector::PathSelector,
     },
 };
